fixed calculation of average value which was performed on long instead of double Added Variance et StandardDeviation to TimerWithMemory
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@1455 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/org/objectweb/proactive/core/util/profiling/Timer.java
@@ -18,7 +18,7 @@ public interface Timer {
 	
 	public int getNumberOfValues();
 	
-	public long getAverage(); 
+	public double getAverage(); 
 	
 	public String getName();
